Back, forward, and list-selection are properly working now, save for a few oversights. Can't claim they are slight - they're actually pretty glaring - but they will be addressed in time as I make the Library Model smarter.
Next objective: first, hook up the scrubber fragment so it knows a song
got paused / started outside of its own control. Also probably give it
some display data on the currently playing song, and update that info
appropriately. Then dig into the profiler, because it's taking a while to
launch, and I'd like to fix that before proceeding.